LayoutTests/editing/spelling/grammar-correction-shown-clicking-middle-of-word.html

<!DOCTYPE html><!-- webkit-test-runner [ spellCheckingDots=true ] -->
<html lang="en">
<meta charset="UTF-8">
<style>
div[contenteditable] {
    font-family: monospace;
    font-size: 32px;
    width: 100%;
}
</style>
<script src="../../resources/ui-helper.js"></script>
<script src="../../resources/js-test.js"></script>

<script>
jsTestIsAsync = true;

description(`This test verifies that grammar correction UI is triggered when clicking
    in the middle of a word with a grammar marker, not just at the end.
    This test requires WebKitTestRunner to run.`);

if (window.internals)
    internals.setContinuousSpellCheckingEnabled(true);

async function runTest()
{
    const editor = document.getElementById("editor");
    const grammarText = "She have a dog";

    await UIHelper.activateElementAndWaitForInputSession(editor);

    document.execCommand("insertText", false, grammarText);
    document.execCommand("insertParagraph");
    await UIHelper.ensurePresentationUpdate();

    // Move selection to start of the text node so hasGrammarMarker checks the right node.
    const selection = window.getSelection();
    selection.collapse(editor.firstChild, 0);

    await shouldBecomeEqual('internals.hasGrammarMarker(4, 4)', 'true');

    // Place cursor in "dog" (a different word) first,
    // then move to the middle of "have".
    // Before the fix, this would not trigger the grammar correction UI
    // because the cursor was not at the end of the word.
    selection.collapse(editor.firstChild, 12);
    await UIHelper.ensurePresentationUpdate();

    selection.collapse(editor.firstChild, 6);
    await shouldBecomeEqual('internals.isAlternativeTextUIActive()', 'true');

    editor.textContent = "";
    editor.blur();
    finishJSTest();
}
</script>
<body>
<div contenteditable="true" id="editor"></div>
<script>
UIHelper.setSpellCheckerResults({
    "She have a dog" : [
        {
            type: "grammar",
            from: 4,
            to: 8,
            details: [{ from: 0, to: 4, corrections: ["has"] }]
        }
    ],
    "She have a dog\n" : [
        {
            type: "grammar",
            from: 4,
            to: 8,
            details: [{ from: 0, to: 4, corrections: ["has"] }]
        }
    ]
}).then(runTest);
</script>
<pre id="description"></pre>
<pre id="console"></pre>
</body>
</html>
